chore: move options cache helper to shared
Relocate options-cache.js from src/plugin/src/options to src/shared/options-cache.js so shared caching utilities live alongside other shared helpers and printer code no longer reaches into the options folder for cross-cutting logic.
