+++ src/plugin/tests/options-cache.test.js
@@ -4,7 +4,7 @@ import test from "node:test";
 import {
     getCachedValue,
     createCachedOptionResolver
-} from "../src/options/options-cache.js";
+} from "../../shared/options-cache.js";
 
 test("getCachedValue caches computed results per options object", () => {
     const options = {};
