+++ src/shared/options-cache.js
@@ -1,4 +1,4 @@
-import { hasOwn, isObjectLike } from "../../../shared/object-utils.js";
+import { hasOwn, isObjectLike } from "./object-utils.js";
 
 const SHARED_CACHE = new WeakMap();
